requirements.txt
absl-py==1.0.0
addict==2.4.0
asgiref==3.4.1
torch==1.10.1
torchvision==0.11.2
cachelib==0.4.1
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
Django==3.2.10
Flask==2.0.2
Flask-Session==0.4.0
future==0.18.2
google-auth==2.3.3
google-auth-oauthlib==0.4.6
grpcio==1.43.0
idna==3.3
imageio==2.13.4
importlib-metadata==4.10.0
itsdangerous==2.0.1
Jinja2==3.0.3
lmdb==1.2.1
Markdown==3.3.6
MarkupSafe==2.0.1
networkx==2.6.3
numpy==1.21.5
oauthlib==3.1.1
opencv-python==4.5.4.60
packaging==21.3
Pillow==8.4.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.6
pytz==2021.3
PyWavelets==1.2.0
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
scikit-image==0.19.1
scipy==1.7.3
six==1.16.0
sqlparse==0.4.2
tb-nightly==2.8.0a20211221
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tifffile==2021.11.2
tqdm==4.62.3
typing_extensions==4.0.1
urllib3==1.26.7
Werkzeug==2.0.2
yapf==0.31.0
zipp==3.6.0
